# openapi/blog.openapi.yaml
openapi: 3.1.0
info:
  title: Drogon Blog API
  version: 1.0.0
  summary: Backend REST surface for the Drogon Blog SPA.
  description: |
    Session-cookie + CSRF-double-submit authentication. Argon2id for
    passwords. Optional 2FA via TOTP, WebAuthn passkeys, and recovery
    codes. PostgreSQL underneath with versioned migrations.

    **Authentication model**

    Login issues an `HttpOnly` `JSESSIONID` session cookie. A
    same-origin request to any state-mutating endpoint must also send
    the `X-CSRF-Token` header set to the value of the `csrf_token`
    cookie (double-submit pattern). Read-only `GET` requests don't
    need the CSRF header.

    **Conditional GETs**

    Listed endpoints return weak ETags (`W/"…"`) and honour
    `If-None-Match` for `304 Not Modified` revalidation. `/auth/me`
    additionally emits `Vary: Cookie` so any shared cache stays
    correctly partitioned by session.

    **Rate limiting**

    Two layers, and they answer different attacks.

    Per-IP token buckets sit in front of `/auth/login`, `/auth/register`,
    `/auth/request-reset`, `/auth/resend-verification`,
    `/auth/reset-password` and `/posts/search`.

    Per-account buckets sit inside the handler for everything that costs
    real work or is an oracle for a secret: `/auth/change-password`, the
    TOTP enrolment pair, `/posts` and `/posts/images` and `/posts/preview`,
    comment creation, message sending, avatar upload, `/reports`, and both
...
    them a fresh budget.

...
servers:
...
security:
  - sessionCookie: []
  - sessionCookie: []
    csrfHeader: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Create a new account
      description: |
        Username and email must be unique. Email is normalised + lightly
        validated; control bytes are rejected to prevent SMTP header
        injection. The password is hashed with Argon2id. A verification
        email is sent asynchronously; the response returns regardless of
        SMTP success so a slow MTA doesn't block the user.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: Username or email already in use.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in with username + password
      description: |
        A fully authenticated success sets the session cookie and the CSRF
        cookie. If the account has 2FA enrolled, the intermediate response
        carries `requires_2fa` and the available methods. Pending login state
        is bound to the rotated, HttpOnly session cookie and expires after ten
        minutes; the client must then complete one of
        `/auth/login/verify-{totp,recovery,webauthn}` with the same cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Either fully authenticated or a 2FA challenge issued.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LoginSuccess'
                  - $ref: '#/components/schemas/TwoStepChallenge'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/AuthDependencyUnavailable' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Sign out
      description: Clears the session and CSRF cookies server-side. Idempotent.
      responses:
        '200':
          description: Logged out.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CsrfMissing' }
  /auth/me:
    get:
      tags: [Auth]
      summary: Current user's identity
      description: |
        Returns the signed-in user's profile. Emits `Vary: Cookie` and a
        weak ETag so a browser back-button / prefetch returns 304
        without leaking another session's data through a shared cache.
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Current user.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Vary:
              schema: { type: string, const: Cookie }
              description: Tells caches to partition by Cookie header.
            Cache-Control:
              schema: { type: string, example: 'private, max-age=0, must-revalidate' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '304': { $ref: '#/components/responses/NotModified' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Consume an email-verification token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, description: Single-use token from the verification email. }
      responses:
        '200':
          description: Email verified.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }
        '400':
          description: Invalid or expired token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/request-reset:
    post:
      tags: [Auth]
      summary: Request a password-reset email
      description: |
        Returns 200 regardless of whether the email exists, so an
        attacker cannot enumerate registered addresses. A reset link is
        sent asynchronously if the address is on file.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: Email queued (always, regardless of account existence).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Consume a reset token and set a new password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:    { type: string }
                password: { type: string, format: password, minLength: 10 }
      responses:
        '200':
          description: Password reset.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }
        '400':
          description: Token invalid / expired / weak password.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /reports:
    post:
      tags: [Moderation]
      summary: Report a post or a comment
      description: |
        Open to any signed-in user. Rate-limited per account, because a
        report queue that anyone can flood is a queue nobody reads.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_type, target_id, reason]
              properties:
                target_type: { type: string, enum: [post, comment] }
                target_id:   { type: integer, minimum: 1 }
                reason:      { type: string }
                detail:      { type: string }
      responses:
        '201':
          description: Filed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  report: { $ref: '#/components/schemas/Report' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CsrfMissing' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }

  # ---------------------------------------------------------- Moderation
  #
  # Every endpoint below answers 404 rather than 403 to a signed-in caller
  # who lacks the role. The moderation surface is not something an ordinary
  # user should be able to map: a 403 confirms the endpoint exists and is
  # worth attacking, while a 404 makes probing it indistinguishable from
  # probing nonsense. A missing session still gets 401 — that is not a
  # secret.
  /admin/reports:
    get:
      tags: [Moderation]
      summary: The report queue
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [open, resolved, dismissed] }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items: { $ref: '#/components/schemas/Report' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/reports/{id}/resolve:
    post:
      tags: [Moderation]
      summary: Close a report
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [resolved, dismissed] }
                note:   { type: string }
      responses:
        '200':
          description: Closed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  report: { $ref: '#/components/schemas/Report' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CsrfMissing' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/posts/{id}/hide:
    post:
      tags: [Moderation]
      summary: Hide a post from every reader
      description: |
        Hiding, not deleting: the row stays so the report it came from still
        points at something, and so the decision can be undone.
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200': { description: Hidden. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CsrfMissing' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/posts/{id}/unhide:
    post:
      tags: [Moderation]
      summary: Restore a hidden post
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200': { description: Restored. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CsrfMissing' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/comments/{id}/hide:
    post:
      tags: [Moderation]
      summary: Hide a comment from every reader
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200': { description: Hidden. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CsrfMissing' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/comments/{id}/unhide:
    post:
      tags: [Moderation]
      summary: Restore a hidden comment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200': { description: Restored. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CsrfMissing' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/users/{id}/ban:
    post:
      tags: [Moderation]
      summary: Suspend an account
      description: |
        A suspension blocks writes immediately, not just new logins — a ban
        that leaves an already-signed-in abuser posting is not a ban. It is
        enforced centrally by a request advice rather than per handler, so a
        new mutating endpoint cannot silently be exempt.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                days:   { type: integer, minimum: 1, default: 7 }
                reason: { type: string }
      responses:
        '200': { description: Suspended. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CsrfMissing' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/users/{id}/unban:
    post:
      tags: [Moderation]
      summary: Lift a suspension
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200': { description: Lifted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/CsrfMissing' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --------------------------------------------------------- Feature flags
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: HttpOnly session cookie issued by `/auth/login`.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Double-submit value. Read its expected value from the
        non-HttpOnly `csrf_token` cookie and echo it back on every
        state-mutating request. Required on `POST`/`PUT`/`PATCH`/`DELETE`
        (except `/auth/login` and `/auth/register` which bootstrap the
        cookie).
    bearerMetricsToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Set via the `BLOG_METRICS_TOKEN` env var. Optional in dev.
  parameters:
    IfNoneMatch:
      in: header
      name: If-None-Match
      required: false
      schema: { type: string }
      description: Echo the previous ETag here to revalidate; receive 304 on cache hit.
    PostId:
      in: path
      name: id
      required: true
      schema: { type: integer, minimum: 1 }
  headers:
    ETag:
      schema: { type: string, example: 'W/"6dfbb62472e910b2"' }
      description: Weak ETag (RFC 7232).
  responses:
    NotModified:
      description: Cached copy is fresh. Empty body.
      headers:
        ETag: { $ref: '#/components/headers/ETag' }
    BadRequest:
      description: Malformed request body / invalid parameters.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid session.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    CsrfMissing:
      description: CSRF cookie/header pair missing or mismatched.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    RateLimited:
      description: Per-IP token-bucket budget exhausted.
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    AuthDependencyUnavailable:
      description: |
        A critical authentication dependency is unavailable. No authenticated
        session was issued; the client may retry after the indicated delay.
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1, example: 5 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error description. Stable identifiers are not promised; rely on the HTTP status for branching.
    Message:
      type: object
      required: [message]
      properties:
        message: { type: string }
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 64 }
        email:    { type: string, format: email }
        password: { type: string, format: password, minLength: 10 }
    RegisterResponse:
      type: object
      required: [message]
      properties:
        message:    { type: string }
        email_sent: { type: boolean }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    LoginSuccess:
      type: object
      required: [message, user]
      properties:
        message: { type: string, const: Login successful }
        user:    { $ref: '#/components/schemas/CurrentUser' }
    TwoStepChallenge:
      type: object
      required: [requires_2fa, methods]
      properties:
        requires_2fa: { type: boolean, const: true }
        methods:
          type: array
          minItems: 1
          items: { type: string, enum: [totp, recovery, webauthn] }
...
    CurrentUser:
      type: object
      required: [id, username, email]
      properties:
        id:       { type: integer, minimum: 1 }
        username: { type: string }
        email:    { type: string, format: email }
        bio:      { type: string }
        profile_image: { type: string }
        role:
          type: string
          enum: [user, moderator, admin]
          description: |
            The caller's own role, read fresh from the database on each
            request. Clients use it only to decide whether to show the
            moderation UI; every `/admin/*` endpoint re-checks the role
            server-side and answers 404 to non-staff. Omitted only if the
            role lookup itself failed.
...
    Report:
      type: object
      required: [id, target_type, target_id, reason, status, created_at]
      properties:
        id:          { type: integer }
        target_type: { type: string, enum: [post, comment] }
        target_id:   { type: integer }
        reason:      { type: string }
        detail:      { type: string }
        status:      { type: string, enum: [open, resolved, dismissed] }
        created_at:  { type: string, format: date-time }
        resolved_at: { type: string, format: date-time, nullable: true }
        resolution_note: { type: string, nullable: true }
...